+++ src/datatypes/vec3.c
@@ -6,23 +6,23 @@
 /*   By: jcat <joaoteix@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/31 19:44:26 by jcat              #+#    #+#             */
-/*   Updated: 2024/04/18 16:03:38 by psotto-m         ###   ########.fr       */
+/*   Updated: 2024/05/05 16:42:16 by joaoteix         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vec3.h"
 
-float	length_squared(t_vec3 *v)
+double	length_squared(t_vec3 *v)
 {
 	return (v->x * v->x + v->y * v->y + v->z * v->z);
 }
 
-float	v3length(t_vec3 *v)
+double	v3length(t_vec3 *v)
 {
 	return (sqrt(length_squared(v)));
 }
 
-float	v3dot(const t_vec3 u, const t_vec3 v)
+double	v3dot(const t_vec3 u, const t_vec3 v)
 {
 	return (u.x * v.x
 		+ u.y * v.y
@@ -38,7 +38,7 @@ t_vec3	v3cross(const t_vec3 u, const t_vec3 v)
 
 t_vec3	v3unit(t_vec3 v)
 {
-	const float	len = v3length(&v);
+	const double	len = v3length(&v);
 
 	return (v3scalef(v, 1.0f / (len + (len == 0.f))));
 }
